refactor(rust): rename `Finalizer` to `ScopeHoistingFinalizer`
