+++ crates/rolldown/src/module_finalizers/scope_hoisting/impl_visit_mut.rs
@@ -14,9 +14,9 @@ use rolldown_oxc_utils::{ExpressionExt, IntoIn, StatementExt, TakeIn};
 
 use crate::utils::call_expression_ext::CallExpressionExt;
 
-use super::Finalizer;
+use super::ScopeHoistingFinalizer;
 
-impl<'me, 'ast> VisitMut<'ast> for Finalizer<'me, 'ast> {
+impl<'me, 'ast> VisitMut<'ast> for ScopeHoistingFinalizer<'me, 'ast> {
   #[allow(clippy::too_many_lines, clippy::match_same_arms)]
   fn visit_program(&mut self, program: &mut ast::Program<'ast>) {
     let old_body = program.body.take_in(self.alloc);
